move SequentMap
